refactor: rename Booking to Bookings
Change Navigation to components folder
