+++ src/App.jsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import { Home, Booking, BookingList, Overview, Navigation } from './pages';
+import { Home, Bookings, BookingList, Overview } from './pages';
+import { Navigation } from './components';
 import './App.css';
 
 function App() {
@@ -11,7 +12,7 @@ function App() {
          <Navigation />
          <Routes>
             <Route path="/" element={<Home />} />
-            <Route path="/bookings" element={<Booking bookings={bookings} setBookings={setBookings} />} />
+            <Route path="/bookings" element={<Bookings bookings={bookings} setBookings={setBookings} />} />
             <Route path="/overview" element={<Overview />} />
             <Route path="/booking-list" element={<BookingList bookings={bookings} />} />
          </Routes>
