+++ src/components/index.js
@@ -2,4 +2,5 @@ export { default as Button } from './ButtonDel';
 export { default as Inventory } from './Inventory';
 export { default as MaterialLab } from './MaterialLab';
 export { default as MaterialTable } from './MaterialTable';
+export { default as Navigation } from './Navigation';
 export { default as SaveData } from './SaveData';
