+++ src/pages/index.js
@@ -1,5 +1,4 @@
-export { default as Booking } from './Booking';
+export { default as Bookings } from './Bookings';
 export { default as BookingList } from './BookingList';
 export { default as Home } from './Home';
-export { default as Navigation } from './Navigation';
 export { default as Overview } from './Overview';
